Add flexible datagridview and all 10 films

+++ FilmsToWatch/Film.cs
@@ -17,6 +17,6 @@ namespace FilmsToWatch
         private string productionCompany;
         private DateTime releasedDate;
         private string language;
-        private ulong budget;
+        private decimal budget;
     }
 }
